refactor: inline validate_config
