+++ modules/fedimint-dummy-common/src/config.rs
@@ -1,11 +1,10 @@
-use anyhow::bail;
 use fedimint_core::config::{
     TypedClientModuleConfig, TypedServerModuleConfig, TypedServerModuleConsensusConfig,
 };
 use fedimint_core::core::ModuleKind;
 use fedimint_core::encoding::{Decodable, Encodable};
 use fedimint_core::module::ModuleConsensusVersion;
-use fedimint_core::{Amount, PeerId};
+use fedimint_core::Amount;
 use serde::{Deserialize, Serialize};
 use threshold_crypto::serde_impl::SerdeSecret;
 use threshold_crypto::{PublicKeySet, SecretKeyShare};
@@ -67,18 +66,6 @@ impl TypedServerModuleConfig for DummyConfig {
     // TODO: Boilerplate-code (remove local)
     fn to_parts(self) -> (ModuleKind, Self::Local, Self::Private, Self::Consensus) {
         (KIND, (), self.private, self.consensus)
-    }
-
-    /// Checks whether the config should be used
-    fn validate_config(&self, identity: &PeerId) -> anyhow::Result<()> {
-        let our_id = identity.to_usize();
-        let our_share = self.consensus.public_key_set.public_key_share(our_id);
-
-        // Check our private key matches our public key share
-        if self.private.private_key_share.public_key_share() != our_share {
-            bail!("Private key doesn't match public key share");
-        }
-        Ok(())
     }
 }
