+++ modules/fedimint-mint-common/src/config.rs
@@ -1,14 +1,12 @@
 use std::collections::BTreeMap;
 
-use anyhow::bail;
 use fedimint_core::config::{
     ClientModuleConfig, TypedClientModuleConfig, TypedServerModuleConfig,
     TypedServerModuleConsensusConfig,
 };
 use fedimint_core::core::ModuleKind;
 use fedimint_core::encoding::{Decodable, Encodable};
-use fedimint_core::{Amount, PeerId, Tiered};
-use itertools::Itertools;
+use fedimint_core::{PeerId, Tiered};
 use serde::{Deserialize, Serialize};
 use tbs::{AggregatePublicKey, PublicKeyShare};
 
@@ -84,32 +82,6 @@ impl TypedServerModuleConfig for MintConfig {
 
     fn to_parts(self) -> (ModuleKind, Self::Local, Self::Private, Self::Consensus) {
         (KIND, (), self.private, self.consensus)
-    }
-
-    fn validate_config(&self, identity: &PeerId) -> anyhow::Result<()> {
-        let sks: BTreeMap<Amount, PublicKeyShare> = self
-            .private
-            .tbs_sks
-            .iter()
-            .map(|(amount, sk)| (amount, sk.to_pub_key_share()))
-            .collect();
-        let pks: BTreeMap<Amount, PublicKeyShare> = self
-            .consensus
-            .peer_tbs_pks
-            .get(identity)
-            .unwrap()
-            .as_map()
-            .iter()
-            .map(|(k, v)| (*k, *v))
-            .collect();
-        if sks != pks {
-            bail!("Mint private key doesn't match pubkey share");
-        }
-        if !sks.keys().contains(&Amount::from_msats(1)) {
-            bail!("No msat 1 denomination");
-        }
-
-        Ok(())
     }
 }
